fix warning under VC7; double-to-int conversion.

+++ tu-testbed/gameswf/gameswf_action.cpp
@@ -963,7 +963,10 @@ namespace gameswf
 
 				case 0x24:	// duplicate clip (sprite?)
 				{
-					env->get_target()->clone_display_object(env->top(2).to_tu_string(), env->top(1).to_tu_string(), env->top(0).to_number());
+					env->get_target()->clone_display_object(
+						env->top(2).to_tu_string(),
+						env->top(1).to_tu_string(),
+						(int) env->top(0).to_number());
 					env->drop(3);
 					break;
 				}
